+++ src/test/java/com/mageddo/dnsproxyserver/dnsconfigurator/DnsConfiguratorDefaultTest.java
@@ -1,10 +1,10 @@
 package com.mageddo.dnsproxyserver.dnsconfigurator;
 
 import com.mageddo.dnsproxyserver.templates.IpAddrTemplates;
+import com.sun.jna.Platform;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -13,20 +13,25 @@ import java.util.List;
 
 import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class DnsConfiguratorOSxTest {
+class DnsConfiguratorDefaultTest {
 
   @Spy
   @InjectMocks
-  DnsConfiguratorOSx configurator;
+  DnsConfiguratorDefault configurator;
 
-  @Captor
-  ArgumentCaptor<List<String>> stringListCaptor;
+  @BeforeEach
+  void beforeEach() {
+    assumeTrue(Platform.isWindows() || Platform.isMac());
+  }
 
   @Test
   void mustConfigureDNSServer() {
@@ -54,6 +59,37 @@ class DnsConfiguratorOSxTest {
 
     // assert
     verify(this.configurator).updateDnsServers(eq(network), eq(singletonList(addr.getRawIP())));
+    assertEquals("{WI-FI=[8.8.8.8]}", this.configurator.getServersBefore().toString());
+  }
+
+  @Test
+  void mustStoreServersBeforeOnceAndNotReplaceByOtherValues() {
+    // arrange
+    final var addr = IpAddrTemplates.local();
+    final var network = "WI-FI";
+
+    doReturn(true)
+      .when(this.configurator)
+      .updateDnsServers(any(), any())
+    ;
+
+    doReturn(List.of("8.8.8.8"))
+      .when(this.configurator)
+      .findNetworkDnsServers(eq(network))
+    ;
+
+    doReturn(singletonList(network))
+      .when(this.configurator)
+      .findNetworks()
+    ;
+
+    // act
+    this.configurator.configure(addr);
+    this.configurator.configure(addr);
+
+    // assert
+    verify(this.configurator, times(1)).updateDnsServers(anyString(), any());
+    verify(this.configurator, times(1)).findNetworkDnsServers(any());
     assertEquals("{WI-FI=[8.8.8.8]}", this.configurator.getServersBefore().toString());
   }